Reverted the ^D input commit because it broke resource processing and prompt handling.
Broke subcommands into _do_* _help_* and _complete_* methods.

Renamed the `workspaces add` subcommand to `create`.

Added the `options` command.

Moved the `set` and `unset` commands to subcommands of `options`.

Added the `modules` command.

Moved the `search` and `load` commands to subcommands of `modules`. `search` is now an argument of the `list` subcommand.

Removed the `use` command alias for `load`.

Moved the `show keys` command to the `keys list` subcommand.

Moved the `show modules` command to the `modules list` subcommand.

Moved the `show options` command to the `options list` subcommand.

Moved the `show workspaces` command to the `workspaces list` subcommand.

Created the `goptions` module context command.

Moved the module context `show globals` subcommand to the `list` subcommand of `goptions`